+++ Xendit/M2Invoice/Model/Adminhtml/Source/ChosenMethod.php
@@ -31,10 +31,10 @@ class ChosenMethod implements ArrayInterface
             ['value' => 'dana', 'label' => __('DANA')],
             ['value' => 'indomaret', 'label' => __('Indomaret')],
             ['value' => 'ovo', 'label' => __('OVO')],
-            ['value' => 'linkaja', 'label' => __('LINKAJA')],
+            ['value' => 'linkaja', 'label' => __('LinkAja')],
             ['value' => 'qr_codes', 'label' => __('QR Codes')],
             ['value' => 'dd_bri', 'label' => __('Direct Debit (BRI)')],
-            ['value' => 'kredivo', 'label' => __('Cardless Credit (KREDIVO)')],
+            ['value' => 'kredivo', 'label' => __('Kredivo')],
         ];
 
         if (!$isMultiselect) {
