+++ Xendit/M2Invoice/Model/Payment/AbstractInvoice.php
@@ -244,10 +244,6 @@ class AbstractInvoice extends AbstractMethod
             $chosenMethods = $this->dataHelper->getChosenMethods();
             $currentCode = $this->_code;
 
-            if ($currentCode === 'cchosted') {
-                $currentCode = 'cc';
-            }
-
             if (!in_array($currentCode, explode(',', $chosenMethods))) {
                 return false;
             }
@@ -255,7 +251,7 @@ class AbstractInvoice extends AbstractMethod
 
         $cardPaymentType = $this->dataHelper->getCardPaymentType();
 
-        if (($cardPaymentType === 'popup' && $this->methodCode === 'CCHOSTED') || $this->methodCode === 'CC_SUBSCRIPTION') {
+        if ($this->methodCode === 'CC_SUBSCRIPTION') {
             return true;
         }
