--- Xendit/M2Invoice/Model/Payment/CCHosted.php
@@ -1,330 +1,0 @@
-<?php
-
-namespace Xendit\M2Invoice\Model\Payment;
-
-use Magento\Framework\Phrase;
-use Magento\Payment\Model\InfoInterface;
-use Magento\Framework\Exception\LocalizedException;
-use Magento\Quote\Api\Data\CartInterface;
-
-/**
- * Class CCHosted
- * @package Xendit\M2Invoice\Model\Payment
- */
-class CCHosted extends AbstractInvoice
-{
-    const PLATFORM_NAME = 'MAGENTO2';
-    const CC_HOSTED_PAYMENT_TYPE = 'CREDIT_CARD';
-    /**
-     * Payment code
-     *
-     * @var string
-     */
-    protected $_code = 'cchosted';
-    protected $_minAmount = 10000;
-    protected $_maxAmount = 10000000;
-    protected $_canRefund = true;
-    protected $methodCode = 'CCHOSTED';
-
-    /**
-     * @param InfoInterface $payment
-     * @param float $amount
-     * @return $this|AbstractInvoice
-     * @throws \Magento\Framework\Exception\LocalizedException
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     */
-    public function authorize(InfoInterface $payment, $amount)
-    {
-        $payment->setIsTransactionPending(true);
-
-        $order = $payment->getOrder();
-        $quoteId = $order->getQuoteId();
-        $quote = $this->quoteRepository->get($quoteId);
-
-        if ($quote->getIsMultiShipping()) {
-            return $this;
-        }
-
-        $orderId = $order->getRealOrderId();
-
-        try {
-            $rawAmount = ceil($order->getSubtotal() + $order->getShippingAmount());
-            $billingAddress = $order->getBillingAddress();
-
-            $args = [
-                'order_number'          => $orderId,
-                'amount'                => $amount,
-                'payment_type'          => self::CC_HOSTED_PAYMENT_TYPE,
-                'store_name'            => $this->storeManager->getStore()->getName(),
-                'platform_name'         => self::PLATFORM_NAME,
-                'success_redirect_url'  => $this->dataHelper->getSuccessUrl(false),
-                'failure_redirect_url'  => $this->dataHelper->getFailureUrl($orderId, false)
-            ];
-
-            $promo = $this->calculatePromo($order, $rawAmount);
-
-            if (!empty($promo)) {
-                $args['promotions'] = json_encode($promo);
-                $args['amount']     = $rawAmount;
-
-                $invalidDiscountAmount = $order->getBaseDiscountAmount();
-                $order->setBaseDiscountAmount(0);
-                $order->setBaseGrandTotal($order->getBaseGrandTotal() - $invalidDiscountAmount);
-
-                $invalidDiscountAmount = $order->getDiscountAmount();
-                $order->setDiscountAmount(0);
-                $order->setGrandTotal($order->getGrandTotal() - $invalidDiscountAmount);
-
-                $order->setBaseTotalDue($order->getBaseGrandTotal());
-                $order->setTotalDue($order->getGrandTotal());
-
-                $payment->setBaseAmountOrdered($order->getBaseGrandTotal());
-                $payment->setAmountOrdered($order->getGrandTotal());
-
-                $payment->setAmountAuthorized($order->getGrandTotal());
-                $payment->setBaseAmountAuthorized($order->getBaseGrandTotal());
-            }
-            // send Cardless Credit Payment request
-            $hostedPayment = $this->requestHostedPayment($args);
-
-            if (isset($hostedPayment['error_code'])) {
-                $message = $this->errorHandler->mapInvoiceErrorCode($hostedPayment['error_code']);
-                $this->processFailedPayment($payment, $message);
-            } elseif (isset($hostedPayment['id'])) {
-                $hostedPaymentId = $hostedPayment['id'];
-                $hostedPaymentToken = (isset($hostedPayment['hp_token']))? $hostedPayment['hp_token'] : '';
-                if (isset($hostedPayment['invoice_url'])) {
-                    $redirectUrl = $hostedPayment['invoice_url']."#credit-card";
-                    $payment->setAdditionalInformation('xendit_redirect_url', $redirectUrl);
-                }
-                if ($hostedPaymentId) {
-                    $payment->setAdditionalInformation('xendit_hosted_payment_id', $hostedPaymentId);
-                }
-                if ($hostedPaymentToken) {
-                    $payment->setAdditionalInformation('xendit_hosted_payment_token', $hostedPaymentToken);
-                }
-
-            } else {
-                $message = 'Error connecting to Xendit. Check your API key';
-                $this->processFailedPayment($payment, $message);
-
-                throw new LocalizedException(
-                    new Phrase($message)
-                );
-            }
-        } catch (\Exception $e) {
-            $errorMsg = $e->getMessage();
-            throw new LocalizedException(
-                new Phrase($errorMsg)
-            );
-        }
-
-        return $this;
-    }
-
-    /**
-     * @param InfoInterface $payment
-     * @param float $amount
-     * @return $this|AbstractInvoice
-     * @throws LocalizedException
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     */
-    public function refund(InfoInterface $payment, $amount)
-    {
-        $chargeId = $payment->getParentTransactionId();
-
-        if ($chargeId) {
-            $order = $payment->getOrder();
-            $orderId = $order->getRealOrderId();
-            $canRefundMore = $payment->getCreditmemo()->getInvoice()->canRefund();
-            $isFullRefund = !$canRefundMore &&
-                0 == (double)$order->getBaseTotalOnlineRefunded() + (double)$order->getBaseTotalOfflineRefunded();
-
-            
-            $refundData = [
-                'amount' => $amount,
-                'external_id' => $this->dataHelper->getExternalId($orderId, true)
-            ];
-            $refund = $this->requestRefund($chargeId, $refundData);
-
-            $this->handleRefundResult($payment, $refund, $canRefundMore);
-
-            return $this;
-        } else {
-            throw new LocalizedException(
-                __("Refund not available because there is no capture")
-            );
-        }
-    }
-
-    /**
-     * @param $chargeId
-     * @param $requestData
-     * @return mixed
-     * @throws \Exception
-     */
-    private function requestRefund($chargeId, $requestData)
-    {
-        $refundUrl = $this->dataHelper->getCheckoutUrl() . "/payment/xendit/credit-card/charges/:$chargeId/refunds";
-        $refundMethod = \Zend\Http\Request::METHOD_POST;
-
-        try {
-            $refund = $this->apiHelper->request($refundUrl, $refundMethod, $requestData);
-        } catch (\Exception $e) {
-            throw $e;
-        }
-
-        return $refund;
-    }
-
-    /**
-     * @param $requestData
-     * @return mixed
-     * @throws \Exception
-     */
-    public function requestHostedPayment($requestData)
-    {
-        $hostedPaymentUrl = $this->dataHelper->getCheckoutUrl() . "/payment/xendit/hosted-payments";
-        $hostedPaymentMethod = \Zend\Http\Request::METHOD_POST;
-
-        try {
-            $hostedPayment = $this->apiHelper->request(
-                $hostedPaymentUrl,
-                $hostedPaymentMethod,
-                $requestData
-            );
-        } catch (\Exception $e) {
-            throw $e;
-        }
-
-        return $hostedPayment;
-    }
-
-    /**
-     * @param $payment
-     * @param $message
-     */
-    public function processFailedPayment($payment, $message)
-    {
-        $payment->setAdditionalInformation('xendit_failure_reason', $message);
-    }
-
-    /**
-     * @param $payment
-     * @param $refund
-     * @param $canRefundMore
-     * @throws LocalizedException
-     */
-    private function handleRefundResult($payment, $refund, $canRefundMore)
-    {
-        if (isset($refund['error_code'])) {
-            throw new LocalizedException(
-                __($refund['message'])
-            );
-        }
-
-        if ($refund['status'] == 'FAILED') {
-            throw new LocalizedException(
-                __('Refund failed, please check Xendit dashboard')
-            );
-        }
-
-        $payment->setTransactionId(
-            $refund['id']
-        )->setIsTransactionClosed(
-            1
-        )->setShouldCloseParentTransaction(
-            !$canRefundMore
-        );
-    }
-
-    /**
-     * @param $order
-     * @param $rawAmount
-     * @return array
-     * @throws LocalizedException
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
-     */
-    public function calculatePromo($order, $rawAmount)
-    {
-        $promo = [];
-        $ruleIds = $order->getAppliedRuleIds();
-        $enabledPromotions = $this->dataHelper->getEnabledPromo();
-
-        if (empty($ruleIds) || empty($enabledPromotions)) {
-            return $promo;
-        }
-
-        $ruleIds = explode(',', $ruleIds);
-
-        foreach ($ruleIds as $ruleId) {
-            foreach ($enabledPromotions as $promotion) {
-                if ($promotion['rule_id'] === $ruleId) {
-                    $rule = $this->ruleRepo->getById($ruleId);
-                    $promo[] = $this->constructPromo($rule, $promotion, $rawAmount);
-                }
-            }
-        }
-
-        return $promo;
-    }
-
-    /**
-     * @param $rule
-     * @param $promotion
-     * @param $rawAmount
-     * @return array
-     */
-    private function constructPromo($rule, $promotion, $rawAmount)
-    {
-        $constructedPromo = [
-            'bin_list'        => $promotion['bin_list'],
-            'title'           => $rule->getName(),
-            'promo_reference' => $rule->getRuleId(),
-            'type'            => $this->dataHelper->mapSalesRuleType($rule->getSimpleAction()),
-        ];
-        $rate = $rule->getDiscountAmount();
-
-        switch ($rule->getSimpleAction()) {
-            case 'to_percent':
-                $rate = 1 - ($rule->getDiscountAmount() / 100);
-                break;
-            case 'by_percent':
-                $rate = ($rule->getDiscountAmount() / 100);
-                break;
-            case 'to_fixed':
-                $rate = (int)$rawAmount - $rule->getDiscountAmount();
-                break;
-            case 'by_fixed':
-                $rate = (int)$rule->getDiscountAmount();
-                break;
-        }
-
-        $constructedPromo['rate'] = $rate;
-
-        return $constructedPromo;
-    }
-
-    /**
-     * @param CartInterface|null $quote
-     * @return bool
-     */
-    public function isAvailable(CartInterface $quote = null)
-    {
-        if ($quote === null) {
-            return false;
-        }
-
-        $amount = ceil($quote->getSubtotal() + $quote->getShippingAddress()->getShippingAmount());
-
-        if ($amount < $this->_minAmount || $amount > $this->_maxAmount) {
-            return false;
-        }
-
-        if(!$this->dataHelper->getCcHostedActive()){
-            return false;
-        }
-
-        return true;
-    }
-}
