+++ Xendit/M2Invoice/etc/frontend/di.xml
@@ -7,13 +7,6 @@
             </argument>
         </arguments>
     </type>
-    <type name="Magento\Checkout\Model\CcGenericConfigProvider">
-        <arguments>
-            <argument name="methodCodes" xsi:type="array">
-                <item name="xendit" xsi:type="const">Xendit\M2Invoice\Model\Payment\CC::CODE</item>
-            </argument>
-        </arguments>
-    </type>
     <type name="Xendit\M2Invoice\Block\Info">
         <arguments>
             <argument name="data" xsi:type="array">
