+++ Xendit/M2Invoice/etc/payment.xml
@@ -30,9 +30,6 @@ https://github.com/magento/magento2/blob/2.2.0-rc2.1/app/code/Magento/Multishipp
         <method name='cc'>
             <allow_multiple_address>1</allow_multiple_address>
         </method>
-        <method name='cchosted'>
-            <allow_multiple_address>1</allow_multiple_address>
-        </method>
         <method name='cc_subscription'>
             <allow_multiple_address>1</allow_multiple_address>
         </method>
