--- Xendit/M2Invoice/view/frontend/templates/multishipping/cchosted.phtml
@@ -1,29 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-?>
-<script>
-    require([
-        'uiLayout',
-        'jquery'
-    ], function (layout, $) {
-        $(function () {
-            var paymentMethodData = {
-                method: 'cchosted'
-            };
-            layout([
-                {
-                    component: 'Xendit_M2Invoice/js/view/payment/method-renderer/multishipping/cchosted',
-                    name: 'payment_method_cchosted',
-                    method: paymentMethodData.method,
-                    item: paymentMethodData
-                }
-            ]);
-
-            $('body').trigger('contentUpdated');
-        })
-    })
-</script>
-<!-- ko template: getTemplate() --><!-- /ko -->
