+++ Xendit/M2Invoice/view/frontend/web/js/view/payment/method-renderer.js
@@ -55,10 +55,6 @@ define(
                 component: 'Xendit_M2Invoice/js/view/payment/method-renderer/cc'
             },
             {
-                type: 'cchosted',
-                component: 'Xendit_M2Invoice/js/view/payment/method-renderer/cchosted'
-            },
-            {
                 type: 'cc_subscription',
                 component: 'Xendit_M2Invoice/js/view/payment/method-renderer/cc_subscription'
             },
