+++ Xendit/M2Invoice/view/frontend/web/js/view/payment/method-renderer/alfamart.js
@@ -68,11 +68,6 @@ define(
                     return false;
                 }
 
-                if (totals.grand_total < window.checkoutConfig.payment.alfamart.min_order_amount) {
-                    self.messageContainer.addErrorMessage({'message': 'Xendit doesn\'t support purchases less than Rp 10,000.'});
-                    return false;
-                }
-
                 return true;
             }
         });
