+++ Xendit/M2Invoice/view/frontend/web/js/view/payment/method-renderer/dana.js
@@ -84,10 +84,6 @@ define(
                     return false;
                 }
 
-                if (totals.grand_total < window.checkoutConfig.payment.dana.min_order_amount) {
-                    self.messageContainer.addErrorMessage({'message': 'Xendit doesn\'t support purchases less than Rp 1.'});
-                    return false;
-                }
 
                 return true;
             },
