--- Xendit/M2Invoice/view/frontend/web/js/view/payment/method-renderer/multishipping/cchosted.js
@@ -1,17 +1,0 @@
-define(
-    [
-        'Xendit_M2Invoice/js/view/payment/method-renderer/cchosted'
-    ],
-    function (
-        Component
-    ) {
-        'use strict';
-
-        return Component.extend({
-            defaults: {
-                template: 'Xendit_M2Invoice/payment/multishipping/description',
-                redirectAfterPlaceOrder: false
-            }
-        });
-    }
-);
