+++ Xendit/M2Invoice/view/frontend/web/js/view/payment/method-renderer/multishipping/linkaja.js
@@ -25,7 +25,7 @@ define(
                     var linkajaPhoneNumber = $('#linkaja_linkaja_number').val();
 
                     if (!self.isPhoneNumber(linkajaPhoneNumber)) {
-                        alert('Invalid LINKAJA phone number, please check again');
+                        alert('Invalid LinkAja phone number, please check again');
                         self.isPlaceOrderActionAllowed(true);
                         self.unblock();
                         return;
