--- Xendit/M2Invoice/view/frontend/web/template/payment/cc-form.html
@@ -1,116 +1,0 @@
-<fieldset data-bind="attr: {class: 'fieldset payment items ccard ' + getCode(), id: 'payment_form_' + getCode()}">
-    <legend class="legend">
-        <span><!-- ko i18n: 'Credit Card Information'--><!-- /ko --></span>
-    </legend><br />
-    <div class="field type" data-bind="style: {
-                                            margin: '0px',
-                                            height: '0px'
-                                            }">
-        <div class="control">
-            <ul class="credit-card-types">
-                <!-- ko foreach: {data: getCcAvailableTypesValues(), as: 'item'} -->
-                <li class="item" data-bind="css: {
-                                                 _active: $parent.selectedCardType() == item.value,
-                                                 _inactive: $parent.selectedCardType() != null && $parent.selectedCardType() != item.value
-                                                 } ">
-                </li>
-                <!--/ko-->
-            </ul>
-            <input type="hidden"
-                   name="payment[cc_type]"
-                   class="input-text"
-                   value=""
-                   data-bind="attr: {id: getCode() + '_cc_type', 'data-container': getCode() + '-cc-type'},
-                   value: creditCardType
-                   ">
-        </div>
-    </div>
-    <div class="field number required">
-        <label data-bind="attr: {for: getCode() + '_cc_number'}" class="label">
-            <span><!-- ko i18n: 'Credit Card Number'--><!-- /ko --></span>
-        </label>
-        <div class="control">
-            <input type="text"
-                    autocomplete="off"
-                    name="payment[cc_number]"
-                    class="input-text"
-                    maxlength="16"
-                    value=""
-                    data-bind="attr: {
-                                    autocomplete: off,
-                                    id: getCode() + '_cc_number',
-                                    title: $t('Credit Card Number'),
-                                    'data-container': getCode() + '-cc-number',
-                                    'data-validate': JSON.stringify({'required-number':true})},
-                              enable: isActive($parents),
-                              value: creditCardNumber,
-                              valueUpdate: 'keyup' "/>
-        </div>
-    </div>
-    <div class="field date required" data-bind="attr: {id: getCode() + '_cc_type_exp_div'}">
-        <label data-bind="attr: {for: getCode() + '_expiration'}" class="label">
-            <span><!-- ko i18n: 'Expiration Date'--><!-- /ko --></span>
-        </label>
-        <div class="control">
-            <div class="fields group group-2">
-                <div class="field no-label month">
-                    <div class="control">
-                        <select  name="payment[cc_exp_month]"
-                                 class="select select-month"
-                                 data-bind="attr: {id: getCode() + '_expiration', 'data-container': getCode() + '-cc-month', 'data-validate': JSON.stringify({required:true, 'validate-cc-exp':'#' + getCode() + '_expiration_yr'})},
-                                            enable: isActive($parents),
-                                            options: getCcMonthsValues(),
-                                            optionsValue: 'value',
-                                            optionsText: 'month',
-                                            optionsCaption: $t('Month'),
-                                            value: creditCardExpMonth">
-                        </select>
-                    </div>
-                </div>
-                <div class="field no-label year">
-                    <div class="control">
-                        <select name="payment[cc_exp_year]"
-                                class="select select-year"
-                                data-bind="attr: {id: getCode() + '_expiration_yr', 'data-container': getCode() + '-cc-year', 'data-validate': JSON.stringify({required:true})},
-                                           enable: isActive($parents),
-                                           options: getCcYearsValues(),
-                                           optionsValue: 'value',
-                                           optionsText: 'year',
-                                           optionsCaption: $t('Year'),
-                                           value: creditCardExpYear">
-                        </select>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    <!-- ko if: (hasVerification())-->
-    <div class="field cvv required" data-bind="attr: {id: getCode() + '_cc_type_cvv_div'}">
-        <label data-bind="attr: {for: getCode() + '_cc_cid'}" class="label">
-            <span><!-- ko i18n: 'Card Verification Number'--><!-- /ko --></span>
-        </label>
-        <div class="control _with-tooltip">
-            <input type="password"
-                   autocomplete="off"
-                   class="input-text cvv"
-                   name="payment[cc_cid]"
-                   value=""
-                   maxlength="4"
-                   data-bind="attr: {id: getCode() + '_cc_cid',
-                        title: $t('Card Verification Number'),
-                        'data-container': getCode() + '-cc-cvv',
-                        'data-validate': JSON.stringify({'required-number':true})},
-                        enable: isActive($parents),
-                        value: creditCardVerificationNumber" />
-            <div class="field-tooltip toggle">
-                <span class="field-tooltip-action action-cvv"
-                      tabindex="0"
-                      data-toggle="dropdown"
-                      data-bind="attr: {title: $t('What is this?')}, mageInit: {'dropdown':{'activeClass': '_active'}}">
-                    <span><!-- ko i18n: 'What is this?'--><!-- /ko --></span>
-                </span>
-            </div>
-        </div>
-    </div>
-    <!-- /ko -->
-</fieldset>
