--- Xendit/M2Invoice/view/frontend/web/template/payment/cc.html
@@ -1,36 +1,0 @@
-<div class="payment-method" data-bind="css: {'_active': (getCode() == isChecked())}">
-    <div class="payment-method-title field choice">
-        <input type="radio"
-               name="payment[method]"
-               class="radio"
-               data-bind="attr: {'id': getCode()}, value: getCode(), checked: isChecked, click: selectPaymentMethod, visible: isRadioButtonVisible()"/>
-        <label data-bind="attr: {'for': getCode()}" class="label"><span data-bind="text: getTitle()"></span></label>
-    </div>
-    <div class="payment-method-content">
-        <div class="payment-method-note" data-bind="text: getDescription()"></div>
-        <div data-bind="
-                style: {color: 'red'}
-                ">
-            <span data-bind="text: getTestDescription().prefix, style: { 'font-weight': 'bold' }"></span>
-            <span data-bind="text: getTestDescription().content"></span>
-        </div>
-        <form class="form" id="pay" action="#" name="pay" method="post" >
-            <!-- ko template: 'Xendit_M2Invoice/payment/cc-form' --><!-- /ko -->
-        </form>
-        <div class="actions-toolbar">
-            <div class="primary">
-                <button class="action primary checkout"
-                        type="submit"
-                        data-bind="
-                        click: placeOrder,
-                        attr: {title: $t('Place Order')},
-                        css: {disabled: !isPlaceOrderActionAllowed()},
-                        enable: (getCode() == isChecked())
-                        "
-                        disabled>
-                    <span data-bind="i18n: 'Place Order'"></span>
-                </button>
-            </div>
-        </div>
-    </div>
-</div>
